+++ backend/DTOs/RelatorioKpiDTO.cs
@@ -5,7 +5,7 @@ namespace backend.DTOs
         public int TotalEmpresas { get; set; }
         public int TotalNotasFiscais { get; set; }
         public int TotalItens { get; set; }
-        public double ValorTotalNotas { get; set; }
-        public double TotalImpostos { get; set; }
+        public decimal ValorTotalNotas { get; set; }
+        public decimal TotalImpostos { get; set; }
     }
 }
